+++ app/globals.css
@@ -43,8 +43,8 @@ body {
     --sidebar-accent-foreground: 240 5.9% 10%;
     --sidebar-border: 220 13% 91%;
     --sidebar-ring: 217.2 91.2% 59.8%;
-    --highlight: 329.45deg 79.51% 59.8%;
-    --highlight-secondary: 272.67deg 79.65% 55.69%;
+    --highlight: 200deg 79.51% 59.8%;
+    --highlight-secondary: 252.67deg 79.65% 55.69%;
   }
   .dark {
     --background: 20 14.3% 4.1%;
