+++ components/subscription/subscription-banner.tsx
@@ -7,8 +7,8 @@ export function SubscriptionBanner() {
     <SubscriptionWrapper>
       <div className="flex flex-col items-center justify-between gap-4 p-4 sm:flex-row">
         <div className="flex items-center gap-3">
-          <div className="bg-highlight/10 text-highlight hidden h-10 w-10 flex-shrink-0 items-center justify-center rounded-full sm:flex">
-            <Check className="h-5 w-5" />
+          <div className="bg-highlight/10 hidden h-10 w-10 flex-shrink-0 items-center justify-center rounded-full sm:flex">
+            <Check className="text-highlight-secondary h-5 w-5" />
           </div>
           <div>
             <h3 className="text-sm font-medium sm:text-base">
